bling-dashboard/bling-saas/requirements.txt
streamlit
pandas
duckdb
sqlalchemy
bcrypt
requests
python-dotenv
pytz
plotly
